requirements.txt
pypinksign
pycryptodome
pyOpenSSL
python-dateutil
pyyaml
certifi
urllib3
idna
requests
typer
python-dotenv